+++ src/problems/largest_palindrome_product.rs
@@ -1,10 +1,10 @@
 // https://projecteuler.net/problem=4
 pub const PROBLEM_ID: u16 = 4;
 
-fn get_digits(number: u32) -> Vec<u32> {
-    let mut digits: Vec<u32> = Vec::new();
+fn get_digits(number: u64) -> Vec<u64> {
+    let mut digits: Vec<u64> = Vec::new();
 
-    fn push_digits(number: u32, digits: &mut Vec<u32>) {
+    fn push_digits(number: u64, digits: &mut Vec<u64>) {
         if number >= 10 {
             push_digits(number / 10, digits);
         }
@@ -17,7 +17,7 @@ fn get_digits(number: u32) -> Vec<u32> {
     digits
 }
 
-fn is_palindrome(number: u32) -> bool {
+fn is_palindrome(number: u64) -> bool {
     let digits = get_digits(number);
 
     for i in 0..=digits.len() / 2 {
@@ -29,12 +29,12 @@ fn is_palindrome(number: u32) -> bool {
     true
 }
 
-pub fn run() -> u32 {
-    let mut results: Vec<u32> = Vec::new();
+pub fn run() -> u64 {
+    let mut results: Vec<u64> = Vec::new();
 
-    for i in (100u32..999u32).rev() {
-        for x in (100u32..999u32).rev() {
-            let number: u32 = i * x;
+    for i in (100u64..999u64).rev() {
+        for x in (100u64..999u64).rev() {
+            let number: u64 = i * x;
 
             if is_palindrome(number) {
                 results.push(number);
